+++ src/app/about/People/people.tsx
