+++ src/app/about/People/styles.module.css
@@ -17,7 +17,7 @@
   }
   .header h2 {
     font-size: 2.5rem;
-    color: var(--blue-300);
+    color: var(--brand-400);
     text-transform: uppercase;
   }
   
@@ -46,7 +46,7 @@
     font-weight: 600;
     margin-bottom: 1rem;
     text-transform: uppercase;
-    color: var(--blue-300);
+    color: var(--brand-400);
   }
   .text h1 span {
     color: var(--black);
