+++ src/app/about/page.tsx
@@ -1,9 +1,9 @@
 import React from "react";
-import Hero from "./Hero/page";
-import People from "./People/page";
-import Diverse from "./Diverse/page";
-import Why from "./Why/page";
-import Footer from "../home/Footer/page";
+import Hero from "./Hero/hero";
+import People from "./People/people";
+import Diverse from "./Diverse/diverse";
+import Why from "./Why/why";
+import Footer from "../home/Footer/footer";
 
 export default function About() {
   return (
