+++ src/app/esg/styles.module.css
@@ -47,7 +47,7 @@
 }
 
 .heroActionButton {
-  background: var(--blue-300);
+  background: var(--brand-400);
   color: var(--white);
   padding: 0.5em 2em;
   border-radius: 8px;
@@ -119,7 +119,7 @@
 }
 .textHeader h1 {
   font-size: 2rem;
-  color: var(--blue-300);
+  color: var(--brand-400);
   font-weight: bold;
   margin-bottom: 1.5rem;
 }
@@ -147,7 +147,7 @@
 .textItem hr {
   border: 0;
   height: 2px;
-  background-color: var(--blue-300);
+  background-color: var(--brand-400);
   margin-top: 1rem;
 }
 
@@ -161,7 +161,7 @@
 }
 .textContent h1 {
   font-size: 2rem;
-  color: var(--blue-300);
+  color: var(--brand-400);
   font-weight: bold;
   margin-bottom: 1.5rem;
 }
@@ -181,7 +181,7 @@
 .contentItem h3 {
   font-size: 2rem;
   font-weight: 600;
-  color: var(--blue-300);
+  color: var(--brand-400);
   margin-bottom: 1rem;
   text-transform: uppercase;
 }
