+++ src/app/globals.css
@@ -11,6 +11,11 @@
   --white: rgba(255, 255, 255, 1);
   --blue-100: rgba(50, 132, 255, 0.13);
   --blue-200: rgba(73, 145, 255, 1);
+  --orange-400: #ff7a3d;
+
+  --brand-400: var(--orange-400);
+  --brand-300: var(--orange-400);
+  --brand-200: #ff934f;
 }
 
 * {
