+++ src/app/home/Partners/partner.tsx
