+++ src/app/home/TrackCounter/styles.module.css
@@ -20,7 +20,7 @@
 .number {
   font-size: 2.5rem;
   font-weight: bold;
-  color: var(--blue-300);
+  color: var(--brand-400);
   margin: 10px 0;
 }
