+++ src/app/insights/styles.module.css
@@ -47,7 +47,7 @@
 }
 
 .heroActionButton {
-  background: var(--blue-300);
+  background: var(--brand-400);
   color: var(--white);
   padding: 0.5em 2em;
   border-radius: 8px;
@@ -119,7 +119,7 @@
 }
 .text h1 {
   font-size: 2rem;
-  color: var(--blue-300);
+  color: var(--brand-400);
   font-weight: bold;
   margin-bottom: 1.5rem;
 }
@@ -138,7 +138,7 @@
   padding-top: 2rem;
 }
 .textContent button {
-  background: var(--blue-300);
+  background: var(--brand-400);
   color: var(--white);
   padding: 1em 2em;
   border-radius: 8px;
