+++ src/app/newsletter/form/styles.module.css
@@ -18,7 +18,7 @@
 .formGroup input[type="email"] {
   width: 100%;
   padding: 10px;
-  border: 1px solid var(--blue-300);
+  border: 1px solid var(--brand-400);
   border-radius: 5px;
   font-size: 16px;
 }
@@ -46,7 +46,7 @@
   font-size: 16px;
   font-weight: bold;
   color: #fff;
-  background-color: var(--blue-300);
+  background-color: var(--brand-400);
   border: none;
   border-radius: 4px;
   cursor: pointer;
