+++ src/app/newsletter/styles.module.css
@@ -47,7 +47,7 @@
 }
 
 .heroActionButton {
-  background: var(--blue-300);
+  background: var(--brand-400);
   color: var(--white);
   padding: 0.5em 2em;
   border-radius: 8px;
@@ -119,7 +119,7 @@
 }
 .text h2 {
   font-size: 2rem;
-  color: var(--blue-300);
+  color: var(--brand-400);
   font-weight: bold;
   margin-bottom: 1.5rem;
   text-transform: uppercase;
@@ -142,7 +142,7 @@
   margin-top: 2rem;
 }
 .contactButton button {
-  background: var(--blue-300);
+  background: var(--brand-400);
   color: var(--white);
   padding: 1em 2em;
   border-radius: 8px;
