+++ src/app/strategy/styles.module.css
@@ -47,7 +47,7 @@
 }
 
 .heroActionButton {
-  background: var(--blue-300);
+  background: var(--brand-400);
   color: var(--white);
   padding: 0.5em 2em;
   border-radius: 8px;
@@ -65,7 +65,7 @@
   font-weight: 600;
   font-size: 3rem;
   text-align: center;
-  color: var(--blue-300);
+  color: var(--brand-400);
   text-transform: uppercase;
 }
 .heroTitle span {
@@ -82,7 +82,7 @@
   text-align: center;
 }
 .heroSubtitle h3 span {
-  color: var(--blue-300);
+  color: var(--brand-400);
   text-transform: uppercase;
 }
 .heroSubtitle .listItems {
@@ -90,7 +90,7 @@
   gap: 10px;
   margin: 0;
   padding: 0;
-  color: var(--blue-300);
+  color: var(--brand-400);
   font-size: 1.2rem;
   margin-bottom: 1rem;
 }
@@ -108,7 +108,7 @@
   color: var(--black);
 }
 .heroText button {
-  background-color: var(--blue-300);
+  background-color: var(--brand-400);
   color: var(--white);
   padding: 1rem 3rem;
   border-radius: 8px;
@@ -191,21 +191,21 @@
 }
 
 .cardItems li {
-  display: flex; 
+  display: flex;
   align-items: center;
-  margin: 0.5rem 0; 
+  margin: 0.5rem 0;
   position: relative;
   padding-left: 2.5rem;
 }
 
 .cardItems li::before {
-  content: '';
+  content: "";
   width: 15px;
   height: 15px;
-  background-color: var(--blue-300);
-  border-radius: 50%; 
+  background-color: var(--brand-400);
+  border-radius: 50%;
   position: absolute;
-  left: 0; 
+  left: 0;
   top: 50%;
   transform: translateY(-50%);
 }
@@ -216,7 +216,6 @@
   align-items: center;
   padding: 2rem 6rem;
   background-color: var(--blue-50);
-
 }
 
 .strategy h1 {
@@ -229,8 +228,7 @@
   display: flex;
   flex-wrap: wrap;
   gap: 1rem;
-  justify-content: center; 
-
+  justify-content: center;
 }
 .strategyContainer {
   flex: 1;
@@ -256,23 +254,20 @@
   color: var(--black);
 }
 .divider {
-    height: 2px;
-    background-color: var(--blue-100);
-    border: none;
-    margin: 20px 0;
-  }
-  .funds {
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-    margin: 2rem;  
-  }
-  .funds button {
-    background-color: var(--blue-300);
-    color: var(--white);
-    padding: 1rem 2rem;
-    border-radius: 8px;
-  }
-
-
-
+  height: 2px;
+  background-color: var(--blue-100);
+  border: none;
+  margin: 20px 0;
+}
+.funds {
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+  margin: 2rem;
+}
+.funds button {
+  background-color: var(--brand-400);
+  color: var(--white);
+  padding: 1rem 2rem;
+  border-radius: 8px;
+}
