+++ src/app/syndicate/styles.module.css
@@ -47,7 +47,7 @@
 }
 
 .heroActionButton {
-  background: var(--blue-300);
+  background: var(--brand-400);
   color: var(--white);
   padding: 0.5em 2em;
   border-radius: 8px;
@@ -65,7 +65,7 @@
   font-weight: 600;
   font-size: 3rem;
   text-align: center;
-  color: var(--blue-300);
+  color: var(--brand-400);
   text-transform: uppercase;
 }
 .heroTitle span {
@@ -139,7 +139,7 @@
   text-align: center;
 }
 .button button {
-  background-color: var(--blue-300);
+  background-color: var(--brand-400);
   color: var(--white);
   padding: 1rem 3rem;
   border-radius: 8px;
